requirements.txt
<<<<<<< HEAD
boto3==1.4.4
botocore==1.5.33
=======
awsebcli==3.10.1
blessed==1.14.2
boto3==1.4.4
botocore==1.5.33
cement==2.8.2
codegen==1.0
colorama==0.3.7
>>>>>>> 01500946755e63d74f2b64ba63195098e7cfbe70
defusedxml==0.5.0
Django==1.10.6
django-braces==1.11.0
<<<<<<< HEAD
django-filter==1.0.2
=======
<<<<<<< HEAD
django-rest-auth==0.9.1
=======
>>>>>>> 01500946755e63d74f2b64ba63195098e7cfbe70
django-mako-plus==4.0.2
django-oauth-toolkit==0.12.0
django-rest-auth==0.9.1
django-rest-framework-social-oauth2==1.0.5
django-storages==1.5.2
>>>>>>> database
djangorestframework==3.6.2
<<<<<<< HEAD
djangorestframework-xml==1.3.0
=======
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
>>>>>>> 01500946755e63d74f2b64ba63195098e7cfbe70
docutils==0.13.1
jmespath==0.9.2
Mako==1.0.6
MarkupSafe==1.0
nap==2.0.0
oauthlib==2.0.1
<<<<<<< HEAD
psycopg2==2.7.1
PyJWT==1.4.2
python-dateutil==2.6.0
python-social-auth==0.3.6
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
=======
pathspec==0.5.0
patterns==0.3
psycopg2==2.7.1
PyJWT==1.4.2
<<<<<<< HEAD
python3-openid==3.1.0
requests==2.13.0
six==1.10.0
click==6.7
djangorestframework-xml==1.3.0
get==0.0.0
Jinja2==2.9.5
livereload==2.5.1
Markdown==2.6.8
MarkupSafe==1.0
mkdocs==0.16.2
oauthlib==2.0.2
post==0.0.0
public==0.0.0
PyYAML==3.12
query-string==0.0.0
request==0.0.0
setupfiles==0.0.0
simplejson==3.10.0
tornado==4.4.2
urlopen==1.0.0
xmljson==0.1.7
=======
python-dateutil==2.6.0
python-social-auth==0.3.6
python3-openid==3.1.0
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.10
semantic-version==2.5.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.40.0
>>>>>>> database
>>>>>>> 01500946755e63d74f2b64ba63195098e7cfbe70
xmltodict==0.10.2
